+++ commands/badappleDot.js
@@ -14,8 +14,22 @@ module.exports = {
                 .setDescription("jumpedFrame per Frame(6-12 recommanded)")
                 .setRequired(true)
                 .setMinValue(1)
+        })
+        .addIntegerOption(sOp => {
+            return sOp
+                .setName("width")
+                .setDescription("Width(24 recommanded on smartphone)")
+                .setRequired(true)
+                .setMinValue(5)
+        })
+        .addIntegerOption(sOp => {
+            return sOp
+                .setName("height")
+                .setDescription("Height(13 recommanded on smartphone)")
+                .setRequired(true)
+                .setMinValue(5)
         }),
     async execute(client, args, argsStr) {
-        playBadApple(args, "⠄⠃⠆⠖⠇⠶⠏⡶⠟⣩⠿⣪⣫⣾⣿⠏⠟⠿", argsStr[0]);
+        playBadApple(args, "⠄⠃⠆⠖⠇⠶⠏⡶⠟⣩⠿⣪⣫⣾⣿⠏⠟⠿", argsStr[0], argsStr[1], argsStr[2]);
     },
 };